Feat: Overhaul Promote Tools with Detailed Metrics and Add Randomizer
This commit introduces a complete overhaul of the "Promote Tools" screen, replacing the simple grid with a detailed analytics dashboard. It also adds a new "Randomizer" (Spin Wheel) feature for live show hosts.

The "Promote Tools" screen now provides an in-depth overview of promotion performance, categorized into "Audience Reached," "Discovery Impact," and "Buyers Converted." Each section contains specific metrics with descriptions and "Pro Tips" to help sellers optimize their promotion strategy.

**Key Changes:**

*   **Revamped Promote Tools UI (`PromoteToolsFragment.kt`, `fragment_promote_tools.xml`):**
    *   The previous high-level stats (shows, views, followers) and feature grid have been replaced with a detailed, tabbed layout.
    *   Added a date range filter ("Last 30 days," "Last 3 months," etc.).
    *   Introduced three new metric categories:
        *   **Audience Reached:** Tracks metrics like "Number of Show Boosts," "Impressions," and "Community Boosts."
        *   **Discovery Impact:** Measures user engagement with metrics like "Total Taps and Clicks," "CTR," and "Sustained Watches."
        *   **Buyers Converted:** Shows the financial impact through metrics like "First Time Buyers," "Direct Sales," and "Bids from Promotion."

*   **New UI Components for Promote Tools:**
    *   **`item_promote_metric.xml` & `PromoteMetricAdapter.kt`:** A new `RecyclerView` item and adapter to display individual metrics with a title, value, and description.
    *   **`item_pro_tip.xml`:** A new layout to display contextual "Pro Tips" throughout the analytics screen, offering actionable advice.

*   **Randomizer (Spin Wheel) Feature (`AgoraPublisherActivity.kt`):**
    *   A new "Randomizer" option has been added to the "More" menu in the live show publisher view.
    *   This opens a `randomizer_sheet.xml` bottom sheet, which includes controls to show, hide, shuffle, and spin a wheel.
    *   A `LuckyWheelView` has been integrated into `activity_agora_publisher.xml` to display the spin wheel to the audience.

*   **Data Model Update (`GetMyShowResponse.kt`):**
    *   The data types for product dimensions (`height`, `length`, `width`, `weight`) have been changed from `Int?` to `Double?` to handle fractional values.

+++ app/src/main/java/io/bidswipe/app/network/response/GetMyShowResponse.kt
@@ -89,13 +89,13 @@ data class GetMyShowResponse(
             @SerializedName("flash_sale")
             val flashSale: Boolean?,
             @SerializedName("height")
-            val height: Int?,
+            val height: Double?,
             @SerializedName("id")
             val id: Int?,
             @SerializedName("images")
             val images: List<String?>?,
             @SerializedName("length")
-            val length: Int?,
+            val length: Double?,
             @SerializedName("mail_class")
             val mailClass: String?,
             @SerializedName("pricing")
@@ -125,9 +125,9 @@ data class GetMyShowResponse(
             @SerializedName("variant")
             val variant: Any?,
             @SerializedName("weight")
-            val weight: Int?,
+            val weight: Double?,
             @SerializedName("width")
-            val width: Int?
+            val width: Double?
         ) : Serializable {
 
             fun toLiveShowProduct(): LiveShowModel.Product {
